1.2 groundwork: fp32 guidance, 64-token captions, accents kept, dense-caption fetcher
Guidance in fp32. At scale 4 the extrapolation multiplies a logit difference by
four, which leaves fp16's range and hands softmax an inf; the same MaskGIT
checkpoint produced flat colour fields under autocast on a GPU and ordinary
texture on CPU in fp32. The autoregressive sampler had the identical pattern,
so 1.1's own pictures were affected too.

Captions 32 -> 64 tokens. The whole 1.2 plan rests on dense captions and 32 BPE
tokens cut them off after about twenty words, so we would have paid for detail
the model never saw. Sequence 609 -> 641, roughly 5% more compute.

StripAccents is gone, NFD -> NFC. It made dzien and dzień the same token
sequence, so the model could not be told which one to draw. The image tokenizer
turns out to round-trip accented glyphs fine down to ~24px, so the text
vocabulary was the binding constraint, not the picture.

New fetcher for HF webdataset repos, aimed at caption density rather than image
count: RECAP froze model and compute, changed only captions, and object-presence
accuracy went 80.8 -> 86.2 while the control arm did not move (arXiv:2310.16656).
That is our field with no horse in it. Shards stream to disk, not RAM — 9 GB
each against a 20 GB kernel disk.

+++ train/sample.py
@@ -59,7 +59,13 @@ def generate(model, prefix, cfg, scale, temp, top_k):
     logits = model(prefix, kv=kv, pos=0)[:, -1]
     out = []
     for step in range(cfg.image_len):
-        cond, uncond = logits[:n], logits[n:]
+        # Prowadzenie liczone w fp32, nawet gdy model chodzi w fp16.
+        # Ekstrapolacja mnozy roznice logitow przez skale, wiec przy scale 4
+        # i logitach rzedu 10 wynik siega setek — a fp16 konczy sie na 65504.
+        # Po przekroczeniu zakresu softmax dostaje inf i zwraca degeneracje:
+        # ten sam checkpoint MaskGIT-a dawal plaskie plamy koloru na GPU w
+        # autocascie i normalna teksture na CPU w fp32. Ten sam wzor byl tutaj.
+        cond, uncond = logits[:n].float(), logits[n:].float()
         guided = uncond + scale * (cond - uncond)
 
         guided[:, :lo] = -math.inf          # never emit text or special tokens
